Make config warning responsive
