+++ client/src/route/config/Developer.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+const Developer = () => <></>;
+
+export default Developer;
