+++ client/src/route/config/User.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+const User = () => <></>;
+
+export default User;
